+++ src/ScheduleTable.tsx
@@ -19,6 +19,8 @@ import { useDraggable } from "@dnd-kit/core";
 import { CSS } from "@dnd-kit/utilities";
 import { ComponentProps, Fragment, memo, useCallback } from "react";
 import { useScheduleSetter } from "./ScheduleContext.tsx";
+import { LectureDeletePopover } from "./LectureDeletePopover.tsx";
+import { LectureItemPopover } from "./LectureItemPopover.tsx";
 
 const TIMES = [
   ...Array(18)
@@ -149,54 +151,10 @@ const DraggableSchedule = memo(
   }: { tableId: string; id: string; data: Schedule } & ComponentProps<
     typeof Box
   >) => {
-    const { day, range, room, lecture } = data;
-    const { attributes, setNodeRef, listeners, transform } = useDraggable({
-      id,
-    });
-    const { deleteScheduleItem } = useScheduleSetter();
-
-    const leftIndex = DAY_LABELS.indexOf(day as (typeof DAY_LABELS)[number]);
-    const topIndex = range[0] - 1;
-    const size = range.length;
-
     return (
       <Popover>
-        <PopoverTrigger>
-          <Box
-            position="absolute"
-            left={`${120 + CellSize.WIDTH * leftIndex + 1}px`}
-            top={`${40 + (topIndex * CellSize.HEIGHT + 1)}px`}
-            width={CellSize.WIDTH - 1 + "px"}
-            height={CellSize.HEIGHT * size - 1 + "px"}
-            bg={bg}
-            p={1}
-            boxSizing="border-box"
-            cursor="pointer"
-            ref={setNodeRef}
-            transform={CSS.Translate.toString(transform)}
-            {...listeners}
-            {...attributes}
-          >
-            <Text fontSize="sm" fontWeight="bold">
-              {lecture.title}
-            </Text>
-            <Text fontSize="xs">{room}</Text>
-          </Box>
-        </PopoverTrigger>
-        <PopoverContent onClick={(event) => event.stopPropagation()}>
-          <PopoverArrow />
-          <PopoverCloseButton />
-          <PopoverBody>
-            <Text>강의를 삭제하시겠습니까?</Text>
-            <Button
-              colorScheme="red"
-              size="xs"
-              onClick={() => deleteScheduleItem(tableId, data)}
-            >
-              삭제
-            </Button>
-          </PopoverBody>
-        </PopoverContent>
+        <LectureItemPopover data={data} bg={bg} id={id} />
+        <LectureDeletePopover tableId={tableId} data={data} />
       </Popover>
     );
   }
